requirements.txt
Django>=4.1.2
django-cors-headers>=3.13.0
django-filter>=22.1
django-test-without-migrations>=0.6
djangorestframework>=3.14.0
python-decouple>=3.6
psycopg2>=2.9.5
dj-database-url>=1.2.0
gunicorn>=20.1.0
django-extensions>=3.2.1
django-debug-toolbar>=3.8.1
drf-yasg>=1.21.4
dj-rest-auth>=2.2.6
django-allauth>=0.52.0
drf-writable-nested>=0.7.0
django-storages>=1.13.2
boto3>=1.26.60
Pillow>=9.4.0